+++ ftmq/model/dataset.py
@@ -3,17 +3,20 @@ from datetime import datetime
 from typing import Any, ForwardRef, Iterable, Literal, Optional, TypeVar
 
 from nomenklatura.dataset.catalog import DataCatalog as NKCatalog
-from nomenklatura.dataset.coverage import DataCoverage as NKCoverage
 from nomenklatura.dataset.dataset import Dataset as NKDataset
 from nomenklatura.dataset.publisher import DataPublisher as NKPublisher
 from nomenklatura.dataset.resource import DataResource as NKResource
 from normality import slugify
+from pantomime.types import FTM
 from pydantic import AnyUrl, HttpUrl
 
+from ftmq.enums import Categories, Frequencies
 from ftmq.model.coverage import Coverage
 from ftmq.model.mixins import BaseModel, NKModel, RemoteMixin, YamlMixin
+from ftmq.types import CEGenerator
 
-Frequencies = Literal[tuple(NKCoverage.FREQUENCIES)]
+Frequencies = Literal[tuple(Frequencies)]
+Categories = Literal[tuple(Categories)]
 
 C = TypeVar("C", bound="Catalog")
 DS = TypeVar("DS", bound="Dataset")
@@ -28,7 +31,7 @@ class Publisher(NKModel):
     country: str | None = None
     country_label: str | None = None
     official: bool = False
-    logo_uri: AnyUrl | None = None
+    logo_url: HttpUrl | None = None
 
 
 class Resource(NKModel):
@@ -52,7 +55,7 @@ class Maintainer(BaseModel, RemoteMixin, YamlMixin):
     name: str
     description: str | None = None
     url: HttpUrl | None = None
-    logo_uri: HttpUrl | None = None
+    logo_url: HttpUrl | None = None
 
 
 Catalog = ForwardRef("Catalog")
@@ -70,14 +73,18 @@ class Dataset(NKModel):
     url: HttpUrl | None = None
     updated_at: datetime | None = None
     version: str | None = None
-    category: str | None = None
+    category: Categories | None = None
     publisher: Publisher | None = None
     coverage: Coverage | None = None
     resources: list[Resource] | None = []
 
-    # own addition
+    # own addition / aleph
+    countries: list[str] | None = []
+    info_url: HttpUrl | None = None
+    data_url: HttpUrl | None = None
+
     git_repo: AnyUrl | None = None
-    uri: AnyUrl | None = None
+    uri: str | None = None
     maintainer: Maintainer | None = None
     catalog: Optional[Catalog] = None
 
@@ -91,12 +98,19 @@ class Dataset(NKModel):
         )
         data["catalog"] = data.get("catalog") or Catalog().dict()
         super().__init__(**data)
-        self.prefix = data.get("prefix") or slugify(self.name)
+        self.prefix = self.prefix or data.get("prefix") or slugify(self.name)
         self.coverage = self.coverage or Coverage()
         self.title = self.title or self.name.title()
 
     def to_nk(self):
         return self._nk_model(self.catalog.to_nk(), self.dict())
+
+    def iterate(self) -> CEGenerator:
+        from ftmq.io import smart_read_proxies  # FIXME
+
+        for resource in self.resources:
+            if resource.mime_type == FTM:
+                yield from smart_read_proxies(resource.url)
 
 
 Dataset.update_forward_refs()
@@ -110,10 +124,14 @@ class Catalog(NKModel):
 
     # own additions
     name: str | None = "default"
+    title: str | None = "Catalog"
+    description: str | None = None
     maintainer: Maintainer | None = None
+    publisher: Publisher | None = None
     url: HttpUrl | None = None
-    uri: AnyUrl | None = None
-    logo_uri: AnyUrl | None = None
+    uri: str | None = None
+    logo_url: HttpUrl | None = None
+    git_repo: AnyUrl | None = None
     catalogs: list[Catalog] | None = []
 
     def __init__(self, **data):
@@ -135,7 +153,11 @@ class Catalog(NKModel):
             yield from catalog.datasets
 
     def get_scope(self) -> NKDataset:
-        catalog = self.to_nk()
+        # FIXME
+        catalog = self.copy()
+        for ds in catalog.datasets:
+            ds.coverage = None
+        catalog = catalog.to_nk()
         return NKDataset(
             catalog,
             {
@@ -154,7 +176,8 @@ class Catalog(NKModel):
     @property
     def names(self) -> set:
         names = set()
-        names.update(self.to_nk().names)
+        for dataset in self.datasets:
+            names.add(dataset.name)
         for catalog in self.catalogs:
             names.update(catalog.names)
         return names
@@ -162,6 +185,10 @@ class Catalog(NKModel):
     @classmethod
     def from_names(cls, names: Iterable[set]) -> Catalog:
         return cls(datasets=[Dataset(name=n) for n in names])
+
+    def iterate(self) -> CEGenerator:
+        for dataset in self.datasets:
+            yield from dataset.iterate()
 
 
 Catalog.update_forward_refs()