+++ tests/test_util.py
@@ -1,4 +1,6 @@
 import sys
+from collections import defaultdict
+from datetime import datetime
 
 import cloudpickle
 import pytest
@@ -40,7 +42,7 @@ def test_util_str_enum():
 
 
 def test_util_unknown_filters():
-    res = (("country", "de", None), ("name", "alice", None))
+    res = (("country", "de", Comparators.eq), ("name", "alice", Comparators.eq))
     args = ("--country", "de", "--name", "alice")
     assert tuple(util.parse_unknown_filters(args)) == res
     args = ("--country=de", "--name=alice")
@@ -51,7 +53,7 @@ def test_util_unknown_filters():
     assert tuple(util.parse_unknown_filters(args)) == ()
 
     args = ("--country", "de", "--year__gte", "2023")
-    res = (("country", "de", None), ("year", "2023", "gte"))
+    res = (("country", "de", Comparators.eq), ("year", "2023", Comparators.gte))
     assert tuple(util.parse_unknown_filters(args)) == res
 
 
@@ -78,3 +80,20 @@ def test_util_country():
     assert util.get_country_code("Deutschland") == "de"
     assert util.get_country_code("Berlin, Deutschland") == "de"
     assert util.get_country_code("Foo") is None
+
+
+def test_util_clean_dict():
+    assert util.clean_dict(defaultdict(list)) == {}
+    data = defaultdict(list)
+    data["foo"] = [1, 2, 3]
+    data["bar"]
+    assert util.clean_dict(data) == {"foo": [1, 2, 3]}
+    assert util.clean_dict("foo") is None
+
+
+def test_util_get_year():
+    assert util.get_year(None) is None
+    assert util.get_year("2023") == 2023
+    assert util.get_year(2020) == 2020
+    assert util.get_year(datetime.now()) >= 2023
+    assert util.get_year("2000-01") == 2000